refactor segment package
