+++ src/main/java/com/hankcs/hanlp/seg/common/Config.java
@@ -9,7 +9,7 @@
  * This source is subject to the LinrunSpace License. Please contact 上海林原信息科技有限公司 to get more information.
  * </copyright>
  */
-package com.hankcs.hanlp.seg;
+package com.hankcs.hanlp.seg.common;
 
 /**
  * 分词器配置项
