+++ src/main/java/com/hankcs/hanlp/seg/common/wrapper/SegmentWrapper.java
@@ -11,7 +11,7 @@
  */
 package com.hankcs.hanlp.seg.common.wrapper;
 
-import com.hankcs.hanlp.seg.Segment;
+import com.hankcs.hanlp.seg.base.AbstractSegment;
 import com.hankcs.hanlp.seg.common.Term;
 import com.hankcs.hanlp.utility.TextUtility;
 
@@ -27,7 +27,7 @@ import java.util.List;
 public class SegmentWrapper
 {
     BufferedReader br;
-    Segment segment;
+    AbstractSegment segment;
     /**
      * 因为next是单个term出去的，所以在这里做一个记录
      */
@@ -37,7 +37,7 @@ public class SegmentWrapper
      */
     int index;
 
-    public SegmentWrapper(BufferedReader br, Segment segment)
+    public SegmentWrapper(BufferedReader br, AbstractSegment segment)
     {
         this.br = br;
         this.segment = segment;
