+++ src/main/java/com/hankcs/hanlp/tokenizer/IndexTokenizer.java
@@ -12,8 +12,7 @@
 package com.hankcs.hanlp.tokenizer;
 
 import com.hankcs.hanlp.HanLP;
-import com.hankcs.hanlp.seg.Segment;
-import com.hankcs.hanlp.seg.Dijkstra.DijkstraSegment;
+import com.hankcs.hanlp.seg.base.AbstractSegment;
 import com.hankcs.hanlp.seg.common.Term;
 
 import java.util.List;
@@ -27,7 +26,7 @@ public class IndexTokenizer
     /**
      * 预置分词器
      */
-    public static final Segment SEGMENT = HanLP.newSegment().enableIndexMode(true);
+    public static final AbstractSegment SEGMENT = HanLP.newSegment().enableIndexMode(true);
     public static List<Term> segment(String text)
     {
         return SEGMENT.seg(text);
