+++ src/main/java/com/hankcs/hanlp/tokenizer/SpeedTokenizer.java
@@ -12,7 +12,7 @@
 package com.hankcs.hanlp.tokenizer;
 
 import com.hankcs.hanlp.seg.Other.DoubleArrayTrieSegment;
-import com.hankcs.hanlp.seg.Segment;
+import com.hankcs.hanlp.seg.base.AbstractSegment;
 import com.hankcs.hanlp.seg.common.Term;
 
 import java.util.List;
@@ -26,7 +26,7 @@ public class SpeedTokenizer
     /**
      * 预置分词器
      */
-    public static final Segment SEGMENT = new DoubleArrayTrieSegment();
+    public static final AbstractSegment SEGMENT = new DoubleArrayTrieSegment();
     public static List<Term> segment(String text)
     {
         return SEGMENT.seg(text.toCharArray());
