+++ src/test/java/com/hankcs/demo/DemoJapaneseNameRecognition.java
@@ -12,7 +12,7 @@
 package com.hankcs.demo;
 
 import com.hankcs.hanlp.HanLP;
-import com.hankcs.hanlp.seg.Segment;
+import com.hankcs.hanlp.seg.base.AbstractSegment;
 import com.hankcs.hanlp.seg.common.Term;
 
 import java.util.List;
@@ -30,7 +30,7 @@ public class DemoJapaneseNameRecognition
                 "林志玲亮相网友:确定不是波多野结衣？",
                 "龟山千广和近藤公园在龟山公园里喝酒赏花",
         };
-        Segment segment = HanLP.newSegment().enableJapaneseNameRecognize(true);
+        AbstractSegment segment = HanLP.newSegment().enableJapaneseNameRecognize(true);
         for (String sentence : testCase)
         {
             List<Term> termList = segment.seg(sentence);
