+++ src/test/java/com/hankcs/demo/DemoPlaceRecognition.java
@@ -12,7 +12,7 @@
 package com.hankcs.demo;
 
 import com.hankcs.hanlp.HanLP;
-import com.hankcs.hanlp.seg.Segment;
+import com.hankcs.hanlp.seg.base.AbstractSegment;
 import com.hankcs.hanlp.seg.common.Term;
 
 import java.util.List;
@@ -28,7 +28,7 @@ public class DemoPlaceRecognition
         String[] testCase = new String[]{
                 "蓝翔给宁夏固原市彭阳县红河镇黑牛沟村捐赠了挖掘机",
         };
-        Segment segment = HanLP.newSegment().enablePlaceRecognize(true);
+        AbstractSegment segment = HanLP.newSegment().enablePlaceRecognize(true);
         for (String sentence : testCase)
         {
             List<Term> termList = segment.seg(sentence);
