+++ src/test/java/com/hankcs/hanlp/model/perceptron/DemoTrainCWS.java
@@ -14,7 +14,7 @@ public class DemoTrainCWS
         System.out.printf("准确率F1:%.2f\n", result.getAccuracy());
         PerceptronSegmenter segment = new PerceptronSegmenter(result.getModel());
         // 也可以用
-//        Segment segment = new AveragedPerceptronSegment(POS_MODEL_FILE);
+//        AbstractSegment segment = new AveragedPerceptronSegment(POS_MODEL_FILE);
         System.out.println(segment.segment("商品与服务"));
     }
 }
